Change the tempest test interval to 2s.
Always seen tempest test fail with api timeout error. Refer to
http://logs.openstack.org/42/469342/2/check/
gate-zun-devstack-dsvm-docker-sql/02b4650/
logs/apache/zun_api.txt.gz#_2017-05-31_08_29_52_459
But run the tempest in local devstack, the logs show there are
lots of rabbitmq messages and seems the rabbitmq cannot handle
so much message. Checked the code there are lots of check status
called in tempest test.

So this patch propose to change the check status interval to 2s
to reduce the numbers of messages to avoid rabbitmq message stuck.

Change-Id: Ibb41ae552bcce0f685834e13a6c3a5836f657cab