The strategy to fix multi-node gate
In the tempest TestContainer, for e.x [1], we get the container with
model.uuid, then assert it with the desire state. The method
_get_container_state will use docker_client to get container. At that
time, the problem happens, docker client can get only container which
run in master node (the node run tempest), so sometimes it may return
None.

The new strategy is passing docker host get from model (model.hostname)
and pass to docker client. By this way, we can get container in the
specific host.

[1] https://github.com/openstack/zun/blob/master/zun/tests/tempest/api/test_containers.py#L408

Change-Id: I6604889d9fc49a4a828eaf0c4014881791bfd3c7
Implements: blueprint multi-node-ci