Change type of command from string to list
Starting from api version 1.20, the type of parameter 'command'
changed from string to list [1]. This patch updates the tempest
test cases about that and bump the api version to 1.20.

This patch also removes test 'test_add_and_remove_sg_to_container'
because this test is not needed (the APIs for adding/removing
security groups are deprecated).

[1] https://review.openstack.org/#/c/575709/

Depends-On: Iee7d2db97eb274da7b5c6348d97475987db88b56
Change-Id: I09d578173e390e57d165aa58b2437c29b7b0273c